fetch csrf secret from endpoint
