+++ packages/applicant/src/utils/csrfMiddleware/secret.ts
@@ -1,45 +1,19 @@
-import {
-  GetSecretValueCommand,
-  SecretsManagerClient,
-  SecretsManagerClientConfig,
-} from '@aws-sdk/client-secrets-manager';
 import { atou } from './utils';
 import { logger } from '../logger';
 import { IS_PRODUCTION } from '../constants';
 
 const HOST = process.env.HOST;
-const CSRF_SECRET_ARN = process.env.CSRF_SECRET_ARN;
-const AWS_REGION = process.env.AWS_REGION;
-
-const fetchCredentials = async () => {
-  const response = await fetch(`${HOST}/api/aws-local`);
-  const json = await response.json();
-  return json;
-};
-
-let _client: SecretsManagerClient;
-
-const getClient = async () => {
-  if (!_client) {
-    const clientConfig: SecretsManagerClientConfig = {};
-    if (!IS_PRODUCTION) {
-      clientConfig.credentials = await fetchCredentials();
-      clientConfig.region = AWS_REGION;
-    }
-    _client = new SecretsManagerClient(clientConfig);
-  }
-  return _client;
-};
+const hostDomain = IS_PRODUCTION
+  ? 'https://localhost:3000/apply/applicant'
+  : HOST;
 
 const fetchSecret = async () => {
-  const getSecretCommand = new GetSecretValueCommand({
-    SecretId: CSRF_SECRET_ARN,
-  });
   try {
     logger.info('Fetching CSRF secret');
-    const client = await getClient();
-    const response = await client.send(getSecretCommand);
-    return atou(JSON.parse(response.SecretString).csrfSecret);
+    const response = await fetch(`${hostDomain}/api/csrf`);
+    logger.http('response', response);
+    const data = await response.json();
+    return atou(data.secret);
   } catch (e) {
     // the aws sdk throws a string, not an error...
     if (typeof e === 'string') throw new Error(e);
